requirements.txt
numpy>=1.13.3
pandas==0.23.3
pafy>=0.5.3.1
youtube-dl>=2017.12.2
pprint>=0.1
hprose>=1.4.3
argparse>=1.4.0
httplib2>=0.10.3
google-api-python-client>=1.6.4
opencv-python
joblib==0.12.0
tensorflow
keras
google-cloud-vision==0.32.0
tqdm==4.23.4
editdistance==0.4
imgaug==0.2.6
h5py
Pillow
cython
